requirements.txt
aiohttp==3.5.4
async-timeout==3.0.1
attrs==19.1.0
beautifulsoup4==4.7.1
bs4==0.0.1
certifi==2019.3.9
cffi==1.12.2
chardet==3.0.4
discord.py==1.2.2
idna==2.8
motor==2.0.0
multidict==4.5.2
pycparser==2.19
pymongo==3.7.2
PyNaCl==1.2.1
redis==3.2.1
requests==2.21.0
six==1.12.0
soupsieve==1.9
uritools==2.2.0
urllib3==1.24.1
websockets==6.0
yarl==1.3.0
youtube-dl==2019.6.8
